renamed screenExtension namespace to screens
